New IDV plugin that interfaces to Unidata's IDV. Very basic for now. Must add the dlls in common/ to java.library.path or have a proper j3d installation.
The KBoxManager is the kbox of all kboxes - queries to it will be dispatched to all.

New IGazetteer interface and bare implementation in geospace. The plugin can register gazetteers and has a lookupFeature method that returns zero or more shapevalues for them. The 'locate' command interfaces to it, if called without parameters it will list known locations.

Starting to restructure the idv command to allow much more with much less effort - including using the gazetteers for space resolution and linking directly to IDV to visualize results. Before that I must add support to visualize categorical and distribution states.

git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/thinklab@2431 ca20354d-2f28-0410-8711-9259e250f5a2