More caching and serialization stuff. Still quite a bit of frustration with it, but it may work at this point once I make the right datasources serializable. Must move to GUI and marine models ASAP.
Added a command line non-graphical shell linked to the commandline.graphical.shell=false|true in the properties of the commandline plugin. Doesn't work on Win at this point, probably because of classpath issues with the dll it generates, but may work on linux where it counts - it's just calling stty there.

git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/thinklab@2601 ca20354d-2f28-0410-8711-9259e250f5a2