Adding lineage tracker and interfaces to support it.
Adding IEntifiable interface to tag those objects that have a peer in the physical world so that their mapping onto it can be recognized and reconstructed.

Made ShapeExtents entifiable and added them to the lineage of grid extents.

Added models to the metadata of observations generated by models.

Fixes and improvements.

git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/thinklab@2709 ca20354d-2f28-0410-8711-9259e250f5a2