Move name generation to UUIDs, add a getSessionWorkspace method to create a valid file name for workspaces given that identifiers now start with urn:uuid:
Create new ArrayMultidimensionalArray that wraps a standard array and has a reduce() method to aggregate along a dimension at a time, waiting to be integrated with states and contexts.

Creation of rankings and measurements with single value added to ObservationFactory in light of integration of parameters with observations. 

Contexts passed to createState() in observations and createState is now called systematically even for POD types, so that metadata in states are under control of  observations in one place.

Waiting to reconceptualize mediated observations as single contingencies and implementing contingency activation handling, so that the whole observation stuff is complete.

git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/thinklab@2724 ca20354d-2f28-0410-8711-9259e250f5a2