Scenario mechanism implemented, first scenario test working. New option in model command allows to apply scenarios:
model -s aries/test/eroded -d aries/view/altitude chehalis

Ready to be tested on deeper hierarchies.

Added a few of many possible clojure bridges to distributions in the ssj package, and ensured that a distribution can be passed as a value to a measurement (remains to be tested on a ranking), so that we can not only write

(measurement 'geophysics:Altitude "m" :value 120.0)

but also, e.g.

(measurement 'geophysics:Altitude "m" :value (corescience/gaussian 120.0 5.4))

which works as expected, sampling the distribution at each context state.


git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/thinklab@2752 ca20354d-2f28-0410-8711-9259e250f5a2