Cleanup of model COW pattern and proper attribution of formal names in dependencies (hopefully). Fixed most, hopefully all, problematic models where :as didn't work and strange ids were needed. If some remain, we'll get nulls so beware.
Models are not "bindings" anymore but get read by the content loader installed by the modelling plugin. As a result all models in models/ contexts/ scenarios/ and agents/ directories are read - meaning that "incubator" models should go somewhere else to avoid damage.

All those objects are stored in the gigantic ModelMap which allows reconstructing the source code for all of them and syncing model changes with their code. This is in preparation for model injection and syncing from external files (read Genie).

Stubs for gazetteer and storyline commands. 

git-svn-id: https://imt.svn.sourceforge.net/svnroot/imt/trunk/thinklab@3517 ca20354d-2f28-0410-8711-9259e250f5a2